Componentise the WYSWIYG functinality
